fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

+++ monasca_agent/collector/checks_d/docker.py
@@ -204,10 +204,10 @@ class Docker(checks.AgentCheck):
                         container_network_dimensions['interface'] = interface_name
                         network_values = cols[1].split()
                         self._report_rate_gauge_metric(
-                            "container.net.in_bytes", long(
+                            "container.net.in_bytes", int(
                                 network_values[0]), container_network_dimensions)
                         self._report_rate_gauge_metric(
-                            "container.net.out_bytes", long(
+                            "container.net.out_bytes", int(
                                 network_values[8]), container_network_dimensions)
                         break
         except Exception as e:
